Likely unnecessary config changes
I made these changes to align with the tutorial I'm following (https://github.com/meech-ward/Bun-Hono-React-Expense-Tracker)

I did it because when trying to deploy to fly.io, the flyctl CLI was "detecting a NodeJS app" instead of a bun app. I thought these changes may help. In the end, it kept saying that, but the dockerfile says bun, so it seems that the CLI message is incorrect but the outcome is right (theory: a bug in flyctl due to bun changing from bun.lockb to bun.lock).

Anyway, not sure if these changes fere necessary, but they seem harmless if not positive, so keeping them.